Added node-based SDF and Shader structure
SDFs and Shaders for models are defined by interconnected nodes.
Currently only Sphere and Lambert are implemented (and Lambert isn't
fully implemented because shading hasn't been implemented into the main
shader). I need to write documentation for the new code but I was making
a lot of progress and wanting to get something working so I will do
that in the next commit