Fixed bug with rendering models
In the branching SDF and shader code the model functions would be called
up to the nearest power of 2 whether they existed or not. This worked
for 1 and 2 models because they are powers of two, but broke on 3
models. This was fixed simply by checking if the index is less than the
model count. It now works correctly and goes only up to the model count.

Also, the FPS is terrible right now. For fullscreen with 3 models it is
running at about 15 FPS. The next update will be optimizations to help
this run faster.