Added max render distance
Raymarching will now stop when the point is past a given distance from
the ray origin. This allows rays to hit the background faster. FPS is
now back up to 60 FPS. The max render distance is specified in glsl.ts
using the paremeter %%max_render_distance%%. Currently it is not
accessible from user code but I will add that later.