Attempted optimization using depth buffer
I tried to optimize it storing a depth buffer of the previous frame to
use as an estimate for the current depth but it doesn't work on the
edges of objects and messes it up.

So my plan now is:

Depth Buffer -> Sobel Edge Detection -> Max Kernel (takes max value in
kernel sample

And then feed this into the main image as well as the depth buffer, and
if it is near an edge (the size of the max kernel will determine how
wide the edges smear to) then just render normally and don't do the
depth optimization because the depth will probably be wrong.

It's going to take a decent amount of restructing and encapsulating shader compilation so I'm commiting now to save this progress